+++ framework/src/main/java/org/checkerframework/common/basetype/BaseTypeChecker.java
@@ -431,8 +431,7 @@ public abstract class BaseTypeChecker extends SourceChecker {
    * @return the type factory of the requested subchecker or null if not found
    */
   @SuppressWarnings("TypeParameterUnusedInFormals") // Intentional abuse
-  public <T extends GenericAnnotatedTypeFactory<?, ?, ?, ?>> @Nullable T getTypeFactoryOfSubchecker(
-      Class<? extends BaseTypeChecker> subCheckerClass) {
+  public <T extends GenericAnnotatedTypeFactory<?, ?, ?, ?>> @Nullable T getTypeFactoryOfSubcheckerOrNull(Class<? extends BaseTypeChecker> subCheckerClass) {
     return getTypeFactory().getTypeFactoryOfSubchecker(subCheckerClass);
   }