+++ framework/src/main/java/org/checkerframework/common/value/ValueCheckerUtils.java
@@ -421,7 +421,7 @@ public class ValueCheckerUtils {
   public static JavaExpression optimize(JavaExpression je, AnnotatedTypeFactory factory) {
     ValueAnnotatedTypeFactory vatf =
         ((GenericAnnotatedTypeFactory<?, ?, ?, ?>) factory)
-            .getTypeFactoryOfSubchecker(ValueChecker.class);
+            .getTypeFactoryOfSubcheckerOrNull(ValueChecker.class);
     return new JavaExpressionOptimizer(vatf == null ? factory : vatf).convert(je);
   }
 }